requirements.txt
smolagents
mcp
mcpadapt
python-dotenv
pandas